extracted debug step, perform interactions, removed safeinputs specific lines and some commented out lines
